Split test host controller implementation
